requirements.txt
asgiref==3.2.10
Django==3.1.2
django-crispy-forms==1.9.2
django-excel==0.0.10
django-model-utils==4.0.0
lml==0.1.0
Pillow==8.0.0
pyexcel==0.6.5
pyexcel-io==0.6.3
pyexcel-webio==0.1.4
pytz==2020.1
sqlparse==0.4.1
texttable==1.6.3
